fix: use float instead of int maxpool

+++ fairseq/modules/funnel_layer.py
@@ -142,7 +142,7 @@ class FunnelEncoderLayer(nn.Module):
     
     def time_compress_encoder_padding_mask(self, mask):
         """Max pool along time axis"""
-        mask = torch.unsqueeze(mask, 0).to(dtype=torch.int32)
+        mask = torch.unsqueeze(mask, 0).to(dtype=torch.float32)
         return self.compress_encoder_padding_mask_fn(mask) > 0
 
     def forward(self, x, encoder_padding_mask, attn_mask: Optional[Tensor] = None):
